requirements.txt
aiohttp==3.7.4.post0
anyio==3.3.0
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.7
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
discord.py==1.7.3
h11==0.12.0
httpcore==0.13.6
httpx==0.19.0
idna==3.2
multidict==5.1.0
pycodestyle==2.7.0
pycparser==2.20
PyNaCl==1.4.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
toml==0.10.2
typing-extensions==3.10.0.0
yarl==1.6.3
youtube-dl==2021.6.6
youtube-search-python==1.4.7
